web/requirements.txt
adrf==0.1.12
aioredis==2.0.1
app-store-server-library==1.9.0
asyncio==4.0.0
boto3==1.42.26
channels==4.3.2
channels-redis==4.3.0
cryptography==46.0.3
dj-database-url==3.1.0
Django==5.2.10
django-stubs==5.2.8
django-stubs-ext==5.2.8
django-allauth==65.13.1
django-cors-headers==4.9.0
django-oauth-toolkit==3.2.0
django-redis==5.4.0
django-storages==1.14.6
djangorestframework==3.16.1
drf-spectacular==0.29.0
gunicorn==23.0.0
hiredis==3.3.0
httpx[http2]==0.28.1
openai==2.15.0
pgvector==0.4.2
psycopg[binary]==3.3.2
pytest==9.0.2
pytest-django==4.11.1
pytest-mock==3.15.1
python-dotenv==1.2.1
pytz==2025.2
redis==7.1.0
sendgrid==6.12.5
sentry-sdk[django]==2.49.0
stripe==14.1.0
taskiq[reload]==0.12.1
taskiq-redis==1.2.1
twilio==9.9.1
uvicorn[standard]==0.40.0
markdown==3.10
langfuse==3.11.2
livekit-api>=1.1.0
